+++ NEWS.md
@@ -1,3 +1,12 @@
+# admiralonco 1.3.0
+
+## Documentation
+
+- New vignette "Creating ADRS with Prostate Cancer Working Group 3 (PCWG3)
+Criteria". (#304)
+
+- Display `ANL01FL` in the `NADIR` derivation in the `ADTR` vignette. (#305)
+
 # admiralonco 1.2.0
 
 ## Documentation